fix(settings): drop redundant u64 cast in debounce_ms test
default_debounce is already u64, so `as u64` is an unnecessary cast
that trips clippy::unnecessary_cast.

+++ src/tui/settings/fields.rs
@@ -1126,7 +1126,7 @@ mod tests {
         );
         assert!(matches!(
             field(&fields, "status_hooks.debounce_ms").value,
-            FieldValue::Number(n) if n == default_debounce as u64
+            FieldValue::Number(n) if n == default_debounce
         ));
 
         let profile = profile_from(json!({"status_hooks": {"debounce_ms": 500}}));
